Lots of compiler (msvc level 4) warning cleanups: [[maybe_unused]] use, and deleted one unused variable, and one variable rename to avoid warning

+++ Library/Sources/Stroika/Foundation/Cryptography/OpenSSL/DerivedKey.cpp
@@ -249,7 +249,7 @@ namespace {
         if (i == 0) {
             Cryptography::OpenSSL::Exception::ThrowLastError ();
         }
-        Assert (i == keyLen);
+        Assert (i == static_cast<int> (keyLen));
         return pair<BLOB, BLOB> (BLOB (useKey.begin (), useKey.end ()), BLOB (useIV.begin (), useIV.end ()));
     }
 }
